+++ src/views/Home.jsx
@@ -4,7 +4,7 @@ export default function Home() {
   return (
     <>
       <h1>Home Page</h1>
-      <Link to="/dashboard">View Dashboard</Link>
+      <Link to="/entries">View Entries</Link>
     </>
   );
 }
